Initial Push of start code
So this is just an amalgamation of all the resources and references needed to build the damn thing. this will change over time obviously.

The first push is the alpha version which i then wanted to try push the beta version over the top of this to see if the beta version will reveal the undocumented changes that differentiate the two versions from alpha to beta.

this is a WIP so dont expect too much at this stage.