Alpha 3 Build
Finally the alpha 3 version is here and has the required and much needed changes added to the code to allow the damn kernel to build with other options.

The proc file system in relation to kcore.c,(about binary files) shm.c (shared memory) , sysctl have all been fixed up and this was the same issues for the beta as well which we will add to that code also to create beta 2 eventually.

thanks to vanvohuyhoang  for his changes that have really saved this project. only took him a few days he has done a great job.