Alpha 3 Source v2
Hello! Finally we have a building version of the original runix beta 2 with the SHM, ksyms and all the other errors have all been fixed up and now you can build the kernel with the default config file and convert it to an ecoff at this stage of release.

i wanna thank @vanvohuyhoang for his work on resolving these issues.

the next step is building the initrd and root fs which will be in the next release.

while i have started wokring on this im still researching the specifics of it all for the playstation end of things.